add task to show routes
