+++ package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "start": "node src/index.js",
     "dev": "nodemon src/index.js",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "routes": "node src/tasks/show-routes.js"
   },
   "repository": {
     "type": "git",
@@ -32,7 +33,8 @@
     "husky": "^7.0.4",
     "nodemon": "^2.0.15",
     "prettier": "^2.6.2",
-    "prisma": "^3.12.0"
+    "prisma": "^3.12.0",
+    "express-list-routes": "^1.1.3"
   },
   "dependencies": {
     "@prisma/client": "^3.12.0",
