+++ src/tasks/show-routes.js
@@ -1,0 +1,5 @@
+import expressListRoutes from 'express-list-routes'
+
+import app from '../app.js'
+
+expressListRoutes(app)
